Improvement: Add cnc|laser shared in pages
